docs/requirements.txt
astropy>=3.2.2
ipython>=7.8.0
matplotlib>=3.1.1
numpy>=1.17.2
plotly>=4.1.1
pytest>=5.2.0
scipy>=1.3.1
tqdm>=4.36.1
wrapt>=1.11.2
sphinx_markdown_tables
numpydoc
sphinx_gallery